move FLAME to separate file
